+++ packages/provider/src/provider.ts
@@ -89,7 +89,7 @@ export function isSequenceProvider(provider: any): provider is Web3Provider {
 
 export class LocalWeb3Provider extends Web3Provider {
   constructor(account: Account, networks?: NetworkConfig[]) {
-    const walletRequestHandler = new WalletRequestHandler(account, null, null, networks || [])
+    const walletRequestHandler = new WalletRequestHandler(account, null, networks || [])
     super(walletRequestHandler)
   }
 }